README.md: fix version
In my previous PR https://github.com/tim-actions/commit-message-checker-with-regex/pull/1
I made a mistake of using `v1.0.0` instead of `v0.1.0` in one place. This PR fixes that.

PS can you please enable "Issues" for this repo? I want to report a problem but can't;
thus I am reporting it below.

-----

-- there is a
tag 'v1' in this repo (https://github.com/tim-actions/commit-message-checker-with-regex/releases/tag/v1)
which I think needs to be removed as the latest version is v0.3.1 and v1 > v0,
so dependabot suggest to update from v0.3.1 to v1 which is obviously wrong.

To fix this, `v1` tag need to be removed.

Alternatively, you can release say v1.1 or so and that will fix the issue
(an update to README.md is required).